+++ SourceGenerator/Tests/CleanPackageTest/CleanTest.Api/GeneratorConfig.cs
@@ -1,0 +1,8 @@
+using REslava.Result.SourceGenerators;
+
+[assembly: GenerateResultExtensions(
+    Namespace = "Generated.ResultExtensions",
+    IncludeErrorTags = true,
+    GenerateHttpMethodExtensions = true,
+    DefaultErrorStatusCode = 400
+)]